[rocprofiler-compute] Split roofline unit tests into test_roofline_main
- Move generate_plot WMMA legend tests and cli_generate_plot guard tests
  out of the analyze integration suite into test_roofline_main.py (1:1
  with roofline_main.py), per PR review.
- Rewrite the cli_generate_plot tests as mock-based: they return at the
  early ai_data/datatype guards and never touched roofline.csv, so the
  workload-dir + sysinfo setup was dead weight.
- Register test_roofline_main in CMakeLists.txt and test_categories.yaml.
