+++ projects/rocprofiler-compute/tests/test_categories.yaml
@@ -21,6 +21,7 @@ test_categories:
       - test_data_imputation
       - test_roofline_calc
       - test_roofline_analyze
+      - test_roofline_main
       - test_num_xcds_spec_class
       - test_profiler_base
       - test_pc_sampling_profile
@@ -76,6 +77,7 @@ test_categories:
       - test_L1_cache_counters
       - test_roofline_calc
       - test_roofline_analyze
+      - test_roofline_main
       - test_torch_trace_analysis
       - test_torch_trace_coverage
       - test_tui_components
@@ -132,6 +134,7 @@ test_categories:
       - test_L1_cache_counters
       - test_roofline_calc
       - test_roofline_analyze
+      - test_roofline_main
       - test_torch_trace_analysis
       - test_torch_trace_coverage
       - test_tui_components
@@ -188,6 +191,7 @@ test_categories:
       - test_L1_cache_counters
       - test_roofline_calc
       - test_roofline_analyze
+      - test_roofline_main
       - test_torch_trace_analysis
       - test_torch_trace_coverage
       - test_tui_components
